Move unit conversion error instrumentation up to DataProvider/namespace data_access to support CSV and NetCDF reporting
